Added ChangeList.
This container is useful to track `Entity` components changes.
It provides some functions to mark `EntityID` as changed and to
mark them as clear again.
It's also possible to fetch the change list, using the function
`for_each`: while fetching, it's possible to safely mark the
`Entities` as updated. In any condition, it does least possible
reallocation while guarantee that all the elements are fetched.

This container doesn't guarantee the order that can change in any
moment.