Exposed changed filter to scripting.
The DynamicQuery got improved, now the code it's more clear and is ready
to support even more filters.

Prepared the DynamicQuery to use the new iterator concept, so the
filters without and maybe are no more determinant for the execution:
meaning that a query can't fetch using only maybe or without (or both)
filters alone, but they always need to be coupled with `with` or
`changed`.

Changed unit tests to make sure the new mechanism is correctly checked
and added test to check the new Changed mechanism is correctly exposed
to scripting.