Improved Query syntax and optimized fetch.
- The static `Query` fetch is optimized, now it iterates only on a given range
of Entities. That `Entities` are the more likely to fulfil the query
requirements, so worth the check.
For example, if you specify a query like this:
`Query<Changed<TransformComponent>>` the `Entities` checked are just the
changed one for that storage, this mean that the fetch is blazing fast.
If you have a more complex query, like this: `Query<TransformComponent, Enemy>`
the query iterates on the storages that have less components; in this
case the `Enemy` storage drives the iteration so again the processing is
super fast.

The `Query` syntax is now improved; thanks to the iterator usage, is now
possible to process the query like this:
```cpp
for(auto [transform, enemy] : query){
	// ...
}
```
The random access is also simplified:
```cpp
if( query.has(1) ){
	auto [transform, enemy] = query[1];
}
```