Added SteadyStorage. Put DenseVectorStorage in its own file
The SteadyStorage is a lot useful when it's necessary to store big
components: in those cases, you don't want to reallocate the memory each time a new
entity is added or removed.

This storage is also useful when you are dealing with libraries like
the physics engine, audio engine, etc... .
With those libraries, it's usually necessary to create objects, so thanks
to this storage it's possible to create components that are fixed in memory,
so pass the pointer or internal component pointers is safe.

Note, the allocated memory is contiguous.