Integrated SharedSteadyStorage.
The SharedSteadyStorage is the perfect choice when you want to share the same component between entities, and that components memory never changes.

When dealing with physics engines or audio engines, etc... it's usually needed to have objects that are shared between some other objects: in all those cases, it's possible to use this storage.