Fully implemented Flatten filter.
With the `Flatten` filter you can specify many components, and it returns
the first valid. The `Flatten` filter, fetches the data if at least one of
its own filters is satisfied.

For example, this `Query<Flatten<TagA, TagB>>` returns all the entities
that contains TagA or TagB.
The syntax to extract the data is the following:
```
Query<Flatten<TagA, TagB>> query;
auto [tag] = query[entity_1];
if( tag.is<TagA>() ){
	TagA* tag_a = tag.as<TagA>();
} else
if( tag.is<TagB>() ){
	TagB* tag_b = tag.as<TagB>();
}
```

Note:
The `Flatten` filter, supports nesting. For example, you can use
the `Changed` filter in this way:
`Query<Flatten<const TagA, Changed<const TagB>>> query;`
Remember that the fist valid filter is returned.
The mutability is also important.

Known limitations:
`Query<Flatten<TagA, TagB>>` if you have an `Entity` that satisfy more
filters, like in the below case (**Entity 2**):
	[Entity 0, TagA, ___]
	[Entity 1, ___, TagB]
	[Entity 2, TagA,TagB]
the query fetches the **Entity 2** twice, but the first specified component
is always taken (in this case the `TagA`).
_Remove this limitation would be a lot more expensinve than useful._