Add Entity node transform synchronization.
Sync the entity transform with the node transform: If active, at the end of the
frame the transform component is copied back to the entity node
transform.

This feature is a lot useful when you move an Entity via system and you
need to also move the node and so the attached childs.