Add the SystemBundle concept.
Implemented the system bundle concept, this is the first commit to move
the pipeline organization from the editor UI to code.

This shift is necessary because the user may not be aware how the systems
must be organized. The developer who implements the systems, will take
care to organize the system order and dependencies. The user will just
need to specify which features needs.