Refactored the ScriptEcs.
The ScriptsEcs class is now used just to handled the scripts and is not
anymore used to access it.

When the script is saved, it's compiled and registered so that it's
possible to access it directly via ECS class.
This allow to fetch the scripted components by ID directly inside the
editor, so that is possible to use the constants: `ECS.MyComponent_gd`
anywhere.

Not only the syntax and usage is improved, but also the performance are
since is no more needed presearch the component by the name.