Finished and improved the pipeline builder.
Now the pipeline builder is able to take into account sub diaptchers.
Was improbed its ability to detect dependencies and sort the systems by implicit depenedencies. Now its able to move systems in batch (when a bundle is found) to fix an explicit dependency. This allow to keep the bundle sorted no matter what.