Refactored the Scripting system and improved DynamicQuery syntax.
Now the DynamicQuery is not implicit in the scripted `System` and the dev has
to specify it inside the `_prepare():` function.

Thanks to this change the internal system code is much simpler and the
system usage is more consistent.

As a side effect, the scripted `System` can have zero or more than one query
(like in C++), giving much more freedom to the dev.

In addition, the `DynamicQuery` got refactored, and it's easier to use.
With the new API it's much easier to fetch all the entities, since you
just need this `while query.next():`; You can also retrieve the
components using dictionary like syntax.

The result code is the following:
```gdscript
extends system

func _prepare():
  var query = Query.new()
  query.with_component(ECS.TransformComponent, MUTABLE)
  with_query(query)

func _execution(query):
  while query.next():
    print(query[0].origin)
    query[0].origin += Vector3(1, 1, 1)
```